refacto player
