--- src/ecs_elements/components/combat.rs
@@ -1,8 +1,0 @@
-
-use bevy::prelude::*;
-use serde::{Serialize, Deserialize};
-
-#[derive(Component, Default, Debug, Serialize, Deserialize, Clone, Copy)]
-pub struct Stats {
-    pub speed: f32
-}
