+++ src/ecs_elements/components/mod.rs
@@ -1,11 +1,6 @@
 mod tiles;
-mod combat;
-mod characters;
 mod movements;
 mod text;
 
-pub use tiles::{TileCollider, TileExit, GameMapRender};
-pub use combat::Stats;
-pub use characters::{Monster, Player, Npc, Piece};
 pub use movements::{Pathfinding, MoveTo, };
 pub use text::AsciiText;
