--- src/ecs_elements/components/tiles.rs
@@ -1,19 +1,0 @@
-use bevy::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::map_builders::commons::TileType;
-
-
-
-
-
-#[derive(Component)]
-pub struct TileCollider;
-
-#[derive(Component)]
-pub struct TileExit;
-
-
-
-#[derive(Component)]
-pub struct GameMapRender;
