+++ src/ecs_elements/resources/mod.rs
@@ -1,7 +1,5 @@
-pub mod save;
 pub mod mapgen;
 pub mod sounds;
 
-pub use save::ShouldSave;
 pub use mapgen::MapGenHistory;
 pub use sounds::MusicController;
