--- src/ecs_elements/resources/save.rs
@@ -1,7 +1,0 @@
-use bevy::prelude::*;
-
-
-#[derive(Resource)]
-pub struct ShouldSave {
-    pub to_save: bool
-}
