+++ src/game/player/components.rs
@@ -1,7 +1,6 @@
 use bevy::prelude::*;
 use serde::{Serialize, Deserialize};
 
-
 #[derive(Component, Default, Debug, Serialize, Deserialize, Clone, Copy)]
 pub struct Player;
 
@@ -13,3 +12,8 @@ pub struct Monster;
 
 #[derive(Component, Default, Debug, Serialize, Deserialize, Clone, Copy)]
 pub struct Piece;
+
+#[derive(Component, Default, Debug, Serialize, Deserialize, Clone, Copy)]
+pub struct Stats {
+    pub speed: f32
+}
