+++ src/render/tilemap_render.rs
@@ -1,11 +1,9 @@
 use bevy::prelude::*;
 
 use crate::{
-    ecs_elements::{
-        components::{TileExit, TileCollider, GameMapRender}
-    }, 
+
     globals::{MAP_WALL, MAP_DEFAULT, MAP_EXIT, MAP_FLOOR},
-    map_builders::TileType, game::{spawners::spawn_sprite_render, GridPosition, Tile}, render::get_world_position
+    map_builders::TileType, game::{spawners::spawn_sprite_render, GridPosition, Tile}, render::{get_world_position, components::{TileCollider, TileExit, GameMapRender}}
 };
